style: move Home styles to a external file
