+++ index.js
@@ -3,7 +3,7 @@
  */
 
 import {AppRegistry} from 'react-native';
-import Home from './src/screens/Home';
+import Home from './src/screens/Onboarding/Home';
 import {name as appName} from './app.json';
 
 AppRegistry.registerComponent(appName, () => Home);
